Fix unit tests for conemu
