+++ reader/src/test/java/org/jline/reader/impl/LineReaderTest.java
@@ -12,21 +12,28 @@ import org.jline.reader.Candidate;
 import org.jline.reader.EndOfFileException;
 import org.jline.reader.LineReader;
 import org.jline.reader.LineReaderBuilder;
+import org.jline.terminal.Size;
 import org.jline.terminal.Terminal;
 import org.jline.terminal.TerminalBuilder;
+import org.jline.terminal.impl.AbstractWindowsTerminal;
 import org.jline.terminal.impl.DumbTerminal;
 import org.jline.utils.AttributedString;
 import org.junit.Ignore;
 import org.junit.Test;
 
+import java.io.BufferedWriter;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.StringWriter;
+import java.nio.charset.Charset;
 import java.util.ArrayList;
 import java.util.List;
 
+import static org.jline.terminal.impl.AbstractWindowsTerminal.TYPE_WINDOWS_CONEMU;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 public class LineReaderTest {
@@ -89,5 +96,50 @@ public class LineReaderTest {
         Terminal terminal = new DumbTerminal(new ByteArrayInputStream(new byte[0]), new ByteArrayOutputStream());
         return new LineReaderImpl(terminal).computePost(c, null, null, "", s -> AttributedString.fromAnsi(s).columnLength(), 80, autoGroup, groupName, true).post.toString();
     }
+
+    @Test
+    public void testConEmuLineReaderClearScreen() throws IOException {
+        System.setProperty("org.jline.terminal.conemu.disable-activate", "false");
+        StringWriter sw = new StringWriter();
+        AbstractWindowsTerminal terminal = new AbstractWindowsTerminal(new BufferedWriter(sw), "name", TYPE_WINDOWS_CONEMU, Charset.defaultCharset(),0,
+                false, Terminal.SignalHandler.SIG_DFL) {
+            @Override
+            protected int getConsoleOutputCP() {
+                return 0;
+            }
+            @Override
+            protected int getConsoleMode() {
+                return 0;
+            }
+            @Override
+            protected void setConsoleMode(int mode) {
+            }
+            @Override
+            protected boolean processConsoleInput() throws IOException {
+                return false;
+            }
+            @Override
+            public Size getSize() {
+                return new Size(80, 25);
+            }
+        };
+        assertTrue(sw.toString().contains("\u001b[9999E"));
+        LineReader reader = new LineReaderImpl(terminal);
+        new Thread(() -> {
+            try {
+                Thread.sleep(50);
+                terminal.processInputChar((char) 12);
+                Thread.sleep(50);
+                terminal.processInputChar('a');
+                terminal.processInputChar((char) 13);
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+        }).start();
+        String line = reader.readLine();
+        assertTrue(sw.toString().contains("\u001b[H\u001b[J"));
+        assertTrue(sw.toString().contains("\u001b[9999E"));
+    }
+
 
 }
